app/pyproject.toml
[tool.poetry]
name = "yandex-queries-telegram-bot"
version = "0.1.0"
description = ""
authors = ["melixz <dr.melix@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.17.0"
aiohttp = "^3.11.11"
python-dotenv = "^1.0.1"
ruff = "^0.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
